third_party/blink/web_tests/platform/mac-retina/virtual/gpu/fast/canvas/OffscreenCanvas-copyImage-expected.txt

This is a testharness.js-based test.
FAIL CopyImageOnTransferredCanvas assert_array_equals: The copied image's top left is red property 0, expected 255 but got 0
Harness: the test ran to completion.
